requirements.txt
gspread==5.7.2
oauth2client==4.1.3
pandas==1.5.2
python_dateutil==2.8.2
requests==2.28.1
